python_backend/services/contacts/pyproject.toml
[tool.poetry]
name = "contacts"
version = "0.1.0"
description = ""
authors = ["Gabriel Garcia <ygabo@pm.me>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
django-extensions = "^3.2.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
faker = "^25.4.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
